services/customer-service/pyproject.toml
[project]
name = "customer-service"
version = "0.1.0"
description = "Customer service MVP for hotel-ddd"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn>=0.34.0,<1.0.0",
  "pydantic>=2.10.0,<3.0.0",
  "pydantic-settings>=2.6.0,<3.0.0",
  "pika>=1.3.2,<2.0.0",
  "sqlalchemy>=2.0.36,<3.0.0",
  "PyJWT>=2.10.0,<3.0.0",
  "bcrypt>=4.2.0,<5.0.0",
  "httpx>=0.28.0,<1.0.0",
  "email-validator>=2.2.0,<3.0.0",
]

[tool.fastapi]
entrypoint = "main:app"

[dependency-groups]
dev = [
  "pytest>=8.3.0,<9.0.0",
  "pytest-cov>=6.0.0,<7.0.0",
  "black>=25.1.0,<26.0.0",
  "ruff>=0.11.0,<0.12.0",
  "pyright>=1.1.0,<2.0.0",
  "testcontainers[rabbitmq]>=4.13.2,<5.0.0",
]

[tool.pytest.ini_options]
testpaths = ["test"]

[tool.black]
line-length = 88

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "B"]

[tool.pyright]
include = ["internal", "main.py"]
venvPath = "."
venv = ".venv"
typeCheckingMode = "basic"
